requirements.txt
Flask==2.2.3
azure-communication-email==1.0.0b1